Split specinfra.rb into specinfra/core.rb.
