+++ lib/specinfra/ext.rb
@@ -1,0 +1,2 @@
+require 'specinfra/ext/class'
+require 'specinfra/ext/string'
