+++ lib/specinfra/ext/class.rb
@@ -1,0 +1,9 @@
+class Class
+  def subclasses
+    result = []
+    ObjectSpace.each_object(Class) do |k|
+      result << k if k < self
+    end
+    result
+  end
+end
